+++ saleor/payment/gateways/stripe/stripe_api.py
@@ -1,24 +1,24 @@
 import json
+import logging
 from decimal import Decimal
+from typing import Optional, Tuple
+from urllib.parse import urljoin
 
 import stripe
 from django.contrib.sites.models import Site
-from typing import List, Optional, Tuple
 from django.urls import reverse
-from stripe.error import StripeError, AuthenticationError
+from stripe.error import AuthenticationError, StripeError
 from stripe.stripe_object import StripeObject
 
-import logging
-from .consts import PLUGIN_ID, WEBHOOK_PATH, WEBHOOK_EVENTS, METADATA_IDENTIFIER
-from urllib.parse import urlunparse, urljoin
 from saleor.core.utils import build_absolute_uri
+
 from ...utils import price_to_minor_unit
-from typing import Callable
+from .consts import METADATA_IDENTIFIER, PLUGIN_ID, WEBHOOK_EVENTS, WEBHOOK_PATH
 
 logger = logging.getLogger(__name__)
 
 
-def is_secret_api_key_valid(api_key:str):
+def is_secret_api_key_valid(api_key: str):
     """Call api to check if api_key is a correct key."""
     try:
         stripe.WebhookEndpoint.list(api_key)
@@ -32,19 +32,19 @@ def subscribe_to_webhook(api_key: str) -> StripeObject:
     api_path = reverse("plugins", kwargs={"plugin_id": PLUGIN_ID})
 
     base_url = build_absolute_uri(api_path)
-    webhook_url = urljoin(base_url, WEBHOOK_PATH)
+    webhook_url = urljoin(base_url, WEBHOOK_PATH)  # type: ignore
 
     return stripe.WebhookEndpoint.create(
         api_key=api_key,
         url=webhook_url,
         enabled_events=WEBHOOK_EVENTS,
-        metadata={
-            METADATA_IDENTIFIER: domain
-        }
+        metadata={METADATA_IDENTIFIER: domain},
     )
 
 
-def change_webhook_status(api_key: str, webhook_id:str, disabled:bool)->Optional[StripeObject]:
+def change_webhook_status(
+    api_key: str, webhook_id: str, disabled: bool
+) -> Optional[StripeObject]:
     try:
         return stripe.WebhookEndpoint.modify(
             webhook_id,
@@ -52,15 +52,19 @@ def change_webhook_status(api_key: str, webhook_id:str, disabled:bool)->Optional
             api_key=api_key,
         )
     except StripeError:
-        logger.warning("Unable to modify a webhook (%s) status to disabled: %s", webhook_id, disabled)
+        logger.warning(
+            "Unable to modify a webhook (%s) status to disabled: %s",
+            webhook_id,
+            disabled,
+        )
         return None
 
 
-def enable_webhook(api_key: str, webhook_id:str) -> StripeObject:
+def enable_webhook(api_key: str, webhook_id: str) -> StripeObject:
     return change_webhook_status(api_key, webhook_id, disabled=False)
 
 
-def disable_webhook(api_key: str, webhook_id:str) -> StripeObject:
+def disable_webhook(api_key: str, webhook_id: str) -> StripeObject:
     return change_webhook_status(api_key, webhook_id, disabled=True)
 
 
@@ -72,21 +76,25 @@ def retrieve_webhook(api_key: str, webhook_id: str) -> Optional[StripeObject]:
         return None
 
 
-def create_payment_intent(api_key: str, amount:Decimal, currency: str) -> Tuple[Optional[StripeObject], Optional[str]]:
+def create_payment_intent(
+    api_key: str, amount: Decimal, currency: str
+) -> Tuple[Optional[StripeObject], Optional[str]]:
     try:
         intent = stripe.PaymentIntent.create(
             api_key=api_key,
             amount=price_to_minor_unit(amount, currency),
-            currency=currency
+            currency=currency,
         )
         return intent, None
     except StripeError as e:
-        error=json.dumps(e.json_body)
+        error = json.dumps(e.json_body)
 
     return None, error
 
 
-def retrieve_payment_intent(api_key: str, payment_intent_id:str)->Optional[StripeObject]:
+def retrieve_payment_intent(
+    api_key: str, payment_intent_id: str
+) -> Optional[StripeObject]:
     try:
         return stripe.PaymentIntent.retrieve(payment_intent_id, api_key=api_key)
     except StripeError: